refactor: move ThreadPool to impl/threading/ (was impl/context/)
context/ was a misleading home — nothing in context/ uses ThreadPool.
impl/threading/ better expresses the intent.
