+++ tt_metal/distributed/dummy_mesh_command_queue.cpp
@@ -4,7 +4,7 @@
 
 #include "dummy_mesh_command_queue.hpp"
 #include <distributed/mesh_device_impl.hpp>
-#include "tt_metal/impl/context/thread_pool.hpp"
+#include "tt_metal/impl/threading/thread_pool.hpp"
 #include <mesh_device.hpp>
 #include <mesh_event.hpp>
 #include <mesh_workload.hpp>
