+++ tt_metal/impl/threading/thread_pool.cpp
@@ -15,7 +15,7 @@
 #include <tt_stl/tt_pause.hpp>
 #include "impl/context/metal_context.hpp"
 #include "impl/context/context_types.hpp"
-#include "impl/context/thread_pool.hpp"
+#include "impl/threading/thread_pool.hpp"
 #include "tt_metal/llrt/tt_cluster.hpp"
 
 namespace tt::tt_metal {
